Allow alternate connections
